Proper handling of login/logout state

+++ weatherfood/views.py
@@ -12,8 +12,9 @@ def index():
 
 @app.route('/weather', methods=['GET','POST'])
 def weather():
-    f = open("username.txt", "r")
-    username = f.read()
+    username = flask.session.get('username', None)
+    if username is None:
+        return flask.redirect('/')
 
     if( api.get_weather_pretty(flask.request.form.get("zipcode")) != "Unavailable" ):
         current_weather = api.get_weather_pretty(
@@ -28,20 +29,24 @@ def weather():
 
 @app.route('/', methods=['GET'])
 def login():
-    return flask.render_template('home.html', url=google.authorization_url())
+    return flask.render_template('home.html', url=google.authorization_url(), hide_logout=True)
 
 @google.login_success
 def login_success(token, profile):
     result = flask.jsonify(token=token, profile=profile)
     username = result.get_json()["profile"]["email"]
     api.create_user(username)
-    f = open("username.txt", "w")
-    f.write("{}".format(username))
+    flask.session['username'] = username
     return flask.render_template('index.html')
 
 @google.login_failure
 def login_failure(e):
     return flask.jsonify(error=str(e))
+
+@app.route('/logout')
+def logout():
+    flask.session['username'] = None
+    return flask.redirect('/')
 
 @app.route('/customrecipe', methods=['POST'])
 def show_recipe():
@@ -53,20 +58,23 @@ def show_recipe():
 
 @app.route('/viewfavorites', methods=['GET'])
 def show_favorites():
-    f = open("username.txt", "r")
-    username = f.read()
+    username = flask.session.get('username', None)
+    if username is None:
+        return flask.redirect('/')
     return flask.render_template( 'favorites.html', favorites = api.retreive_user(username)["Fav_Recipes"] )
 
 @app.route("/del_fav", methods=["GET"])
 def delete_favorite():
-    f = open("username.txt", "r")
-    username = f.read()
+    username = flask.session.get('username', None)
+    if username is None:
+        return flask.redirect('/')
     api.delete_favorite(username, flask.request.values.get('delete'))
     return flask.render_template( 'favorites.html', favorites = api.retreive_user(username)["Fav_Recipes"] )
 
 @app.route('/fav_recipe', methods=['GET'])
 def favorite_recipe():
-    f = open("username.txt", "r")
-    username = f.read()
+    username = flask.session.get('username', None)
+    if username is None:
+        return flask.redirect('/')
     api.update_recipe(username, flask.request.values.get('key'), flask.request.values.get('favorite') )
     return flask.render_template('custom_recipe.html', recipe=temp)
